// src/generated/prisma/models/Product.ts
export type ProductModel = runtime.Types.Result.DefaultSelection<Prisma.$ProductPayload>
// ...
export type $ProductPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Product"
  objects: {
    category: Prisma.$CategoryPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    categoryId: string
    name: string
    slug: string
    description: string | null
    price: runtime.Decimal
    stock: number
    isActive: boolean
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["product"]>
  composites: {}
}